refactor(iter-17): extract tabsHelpers + delete dead-code shims (#352)
architect-expert HIGH (rule 23) + lean-expert MEDIUM (dead code).

* src/store/tabs.ts: 669  519 LOC (-150). Extracted claimOrRevert
  + classifyAndMarkReadOnly into src/store/tabsHelpers.ts (120 LOC,
  new). Trimmed verbose JSDoc on the three Excalidraw maps
  rationale was duplicated across the slice + docs/features/excalidraw.md;
  kept the field-doc one-liners + cross-link.

* src/lib/excalidraw/autosave-banner.ts (27 LOC, deleted) +
  src/lib/excalidraw/first-save-warning.ts (34 LOC, deleted): both
  were 1-line aliases over seenFlag() with a single consumer
  (ExcalidrawView). Inlined the seenFlag calls into ExcalidrawView's
  module scope. Net -49 LOC across the lib.

vitest: 2237/2237 pass (no new tests, no regressions).
lint: 0 errors.
